% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_jags_init}
\alias{create_jags_init}
\title{create_jags_init. Helper function to create jags init list dependant on length on chains provided}
\usage{
create_jags_init(n.chains = 3)
}
\arguments{
\item{n.chains}{Number of chains'}
}
\description{
create_jags_init. Helper function to create jags init list dependant on length on chains provided
}
