% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_BUGS_file}
\alias{extract_BUGS_file}
\title{Helper function to extract BUGS files for given input parameters}
\usage{
extract_BUGS_file(data.type, bth.model, bth.prior, model.pars = NULL,
  model.type = c("PWE", "FP"))
}
\arguments{
\item{data.type}{Only GSD currently supported}

\item{bth.model}{FE or RE}

\item{bth.prior}{list containins priors infomation}

\item{model.pars}{Only needed for FP model, list containing exponets to determin order (1st of 2nd)}

\item{model.type}{Either PWE or FP}
}
\description{
Helper function to extract BUGS files for given input parameters
}
