% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemtcPlus.R
\docType{package}
\name{gemtcPlus}
\alias{gemtcPlus}
\alias{gemtcPlus-package}
\title{gemtcPlus: A package for performing NMA in R}
\description{
This package was designed to work alongside the \code{gemtc} package.
Included are numerous convenience functions to aid in the production of reports.
}
\author{
Sandro Gsteiger \email{sandro.gsteiger@roche.com>}
}
