% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fp_1o.R
\name{get_fp_1o}
\alias{get_fp_1o}
\title{First order fractional polynomial}
\usage{
get_fp_1o(x, params, exponents, sums = NULL)
}
\arguments{
\item{x}{A vector with the dependent variable.}

\item{params}{A matrix with two columns giving the intercept and the "slope". If more than one rows given, the fractional polynomial is evaluated for each row.}

\item{sums}{A function. If non-null, sums of the fractional polynomial at each x value (i.e. summaries over the different params values) are calculated.}

\item{exponent}{A numerical value giving the exponent (p1) of the polynomial.}
}
\value{
A matrix with length(x) rows giving the FP values or FP summaries.
}
\description{
First order fractional polynomial
}
