% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fp_comparison.R
\name{get_fp_comparison}
\alias{get_fp_comparison}
\title{Extract model information and fit statistics from a list of fractional polynomial NMAs.}
\usage{
get_fp_comparison(nmaout.list)
}
\arguments{
\item{nmaout.list}{List of results lists (which must contain the names elements \code{descr_s, model.pars, fit, DICsamp}).}
}
\value{
matrix comparing the models
}
\description{
Extract model information and fit statistics from a list of fractional polynomial NMAs.
}
