% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fp_contrasts.R
\name{get_fp_contrasts}
\alias{get_fp_contrasts}
\title{Extract the treatment contrasts vs the reference in the network}
\usage{
get_fp_contrasts(fit.jg, treatments = NULL, node = "d",
  revert = FALSE)
}
\arguments{
\item{fit.jg}{The jags output object.}

\item{treatments}{A vector with the labels for the treatments in the network (in the correct order). If NULL treatments calculate from fit}

\item{node}{The name of the node identifying the contrast.}
}
\description{
Extract the treatment contrasts vs the reference in the network
}
