% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fp_elements.R
\name{get_fp_elements}
\alias{get_fp_elements}
\title{Extract model information and fit statistics from NMA fit in jags of a fractional polynomial model.}
\usage{
get_fp_elements(nmaout)
}
\arguments{
\item{nmaout}{A list with the results from NMA fit (jags). The list must contain the names elements `descr_s, model.pars, fit, DICsamp`.}
}
\value{
list with description and fit metrics
}
\description{
Extract model information and fit statistics from NMA fit in jags of a fractional polynomial model.
}
