% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fp_S.R
\name{get_fp_S}
\alias{get_fp_S}
\title{Calculate the survivor functions estimated in a fractional polynomial NMA model. The absolute S(t) estimates combining the estimated baseline survival
from a reference trial (in the NMA) with the fractional polynomial (log)hazard ratio estimates to construct the S(t) functions for each treatment.}
\usage{
get_fp_S(fit, ref.std, ref.arm, treatments = NULL, time,
  bl.node = "mu", contrast.node = "d")
}
\arguments{
\item{fit}{JAGS object with NMA fit.}

\item{ref.std}{Numeric identifier of reference study to use for baseline survival estimate.}

\item{ref.arm}{Numeric identifier of arm in reference study to use for baseline survival estimate.}

\item{treatments}{Vector with character strings to label the treatments. If NULL treatments calculated from fit}

\item{time}{Vector of time-points at which S(t) functions are calculated.}

\item{bl.node}{Charactor to identify the node in the jags model that identifies the baseline estimates (default is "mu").}

\item{contrast.node}{Charactor to identify the node in the jags model that identifies the baseline estimates (default is "d").}
}
\value{
A \code{data.frame} containing survivor function for each treatment
}
\description{
Calculate the survivor functions estimated in a fractional polynomial NMA model. The absolute S(t) estimates combining the estimated baseline survival
from a reference trial (in the NMA) with the fractional polynomial (log)hazard ratio estimates to construct the S(t) functions for each treatment.
}
\details{
The FP parameters to calculate the log-hazard curves for each survivor function are combinations \code{bl.node + contrast.node}. This means the combination \code{ref.std, ref.arm} must identify a study arm in the NMA that used the (NMA) reference treatment. The basic parameters (contrasts vs reference) are then added to optain the parameters for each treatment in the study.
}
