% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_jags_info.R
\name{get_jags_info}
\alias{get_jags_info}
\title{Utility function to return jags data and model for reporting (e.g. in appendix)}
\usage{
get_jags_info(sims, include.comments = FALSE,
  input.round.function = function(x) signif(x, digits = 4))
}
\arguments{
\item{sims}{rjags object}

\item{include.comments}{Logical, should comments in the model by included or stripped out (default: FALSE)}

\item{input.round.function}{Function to apply to data inputs - default: function(x) signif(x, digits = 4)}
}
\value{
\code{jagsInfo} object which contains jags simulation information accesed via a call to  \code{print}
}
\description{
Utility function to return jags data and model for reporting (e.g. in appendix)
}
\author{
Iain Bennett (adaptations: Sandro Gsteiger)
}
