% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtc_allVsNew.R
\name{get_mtc_allVsNew}
\alias{get_mtc_allVsNew}
\title{Utility function to extract effect estimates "other treatments vs new" from gemtc fit.}
\usage{
get_mtc_allVsNew(x, new.lab, transform = NULL, digits = NULL)
}
\arguments{
\item{x}{Object of class \code{mtc.result} which is the output from a \code{mtc.run}.}

\item{new.lab}{Character string with name of new intervention.}

\item{transform}{Optional name of transformation to apply to output (e.g. "exp").}

\item{digits}{Optional integer number of digits to round the output to.}
}
\value{
Returns a data.frame of effect estimates
}
\description{
Utility function to extract effect estimates "other treatments vs new" from gemtc fit.
}
