% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtc_probBetter.R
\name{get_mtc_probBetter}
\alias{get_mtc_probBetter}
\title{Utility function to extract probabilities of new treatment being better from gemtc fit (e.g. P(HR<1) for HRs new vs other).}
\usage{
get_mtc_probBetter(x, new.lab, smaller.is.better = TRUE, threshold = 0,
  sort.by = c("name", "effect")[1], digits = 3)
}
\arguments{
\item{x}{Object of class mtc.result containing the NMA fit.}

\item{new.lab}{Character string with name of new intervention.}

\item{smaller.is.better}{Logical (default is TRUE).}

\item{threshold}{Numerical value (default is 0) giving the threshold against which contrasts are being compared.}

\item{sort.by}{Character string, must be either "name" or "effect" (default), indicating the sorting of the output table.}

\item{digits}{Integer (default is 3); if provided, outputs will be rounded accordingly.}
}
\value{
A data.frame of probabilities of a treatment being better from a mtc model
}
\description{
Utility function to extract probabilities of new treatment being better from gemtc fit (e.g. P(HR<1) for HRs new vs other).
}
