% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtc_sum.R
\name{get_mtc_sum}
\alias{get_mtc_sum}
\title{Utility function to extract summary stats from mtc.result object.}
\usage{
get_mtc_sum(x, digits = 2)
}
\arguments{
\item{x}{Object of class \code{mtc.result} which is the output from a \code{mtc.run}.}

\item{digits}{Integer specifiying the number of digits for rounding (default is 2).}
}
\value{
A data.frame of mtc model summary statistics
}
\description{
Utility function to extract summary stats from mtc.result object.
}
