% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nw_fromto.R
\name{get_nw_fromto}
\alias{get_nw_fromto}
\title{Extract edges information ("from-to matrix") from network data frame.}
\usage{
get_nw_fromto(dat)
}
\arguments{
\item{dat}{\code{data.frame} with netowrk data in long format. Must contain columns study and treatment. Can be IPD, grouped data, or AD.}
}
\value{
A \code{matrix} with columns "from", "to".
}
\description{
Extract edges information ("from-to matrix") from network data frame.
}
