% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segments.R
\name{get_pw_segments}
\alias{get_pw_segments}
\title{Utility function to get segments (as character strings) from vector with cutpoints}
\usage{
get_pw_segments(x = NULL, cuts, right = FALSE,
  ordered_results = TRUE)
}
\arguments{
\item{x}{NULL value}

\item{cuts}{Vector with cut points}

\item{right}{Logical}

\item{ordered_results}{Logical}
}
\value{
\code{factor} of segments
}
\description{
Utility function to get segments (as character strings) from vector with cutpoints
}
