% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pwe_conv_diag.R
\name{get_pwe_conv_diag}
\alias{get_pwe_conv_diag}
\title{Utility function: convergence diagnostics for piece-wise constant models}
\usage{
get_pwe_conv_diag(fit, file = NULL, patterns = c("d", "mu"),
  re.pattern = "sd")
}
\arguments{
\item{fit}{\code{rjags} object which is the return of \code{R2jags::jags()}}

\item{file}{Optional output filename (a pdf with this name will be created if provided).}

\item{patterns}{Vector of character strings used to identify the patterns selected for conv assessment.}

\item{re.pattern}{Character}
}
\value{
ggplot object
}
\description{
Utility function: convergence diagnostics for piece-wise constant models
}
