% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segments.R
\name{get_segments}
\alias{get_segments}
\title{Utility function to get segments (as character strings) from vector with cutpoints}
\usage{
get_segments(cut.pts)
}
\arguments{
\item{cut.pts}{Vector with cut points}
}
\value{
\code{factor} of segments
}
\description{
Utility function to get segments (as character strings) from vector with cutpoints
}
