% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupedTTE_fp_pre_proc.R
\name{groupedTTE_fp_pre_proc}
\alias{groupedTTE_fp_pre_proc}
\title{Utility function for pre-processing: prepare jags input data for FP model.}
\usage{
groupedTTE_fp_pre_proc(dat, ref.std = NULL, nma.ref.trt = NULL,
  model.pars = NULL, feprior_mean = NULL, feprior_prec = NULL,
  bth.prior = NULL)
}
\arguments{
\item{dat}{A \code{data.frame} with columns study, treatment, t.start, t.end, n.event, n.risk.}

\item{ref.std}{A \code{character} of the reference study}

\item{nma.ref.trt}{A \code{character} of the reference treatment}

\item{model.pars}{\code{list}, containing exponents (numeric vector of exponents for FP model) and 
t.eval - one of "midpoint" (the default), "start", or "end", to identify the time-point within each interval used to calculate the hazard rate (over each interval, the hazard function is approximated with this value h(t.eval)); alternatively, a  \code{numeric} value in the interval [0,1] used to define t.eval (where for example "midpoint" corresponds to the value 0.5).}

\item{feprior_mean}{A \code{numeric} value representing feprior_mean (default 0)}

\item{feprior_prec}{A \code{numeric} value representing feprior_prec default 0.0001}
}
\value{
A \code{list} with input data for jags fit of PWE model.
}
\description{
Utility function for pre-processing: prepare jags input data for FP model.
}
\seealso{
\code{\link{nma_pre_proc}}, \code{\link{groupedTTE_pwe_pre_proc}}
}
