% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_BUGS}
\alias{list_BUGS}
\title{Lists all available model files inside the inst directory}
\usage{
list_BUGS(full.path = FALSE)
}
\arguments{
\item{full.path}{lgl Return full path to file?}
}
\value{
character vector containing a file name/path
}
\description{
Lists all available model files inside the inst directory
}
