% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{match_args_to_func}
\alias{match_args_to_func}
\title{Helper function to extract named elements from a list to match the arguments of supplied function}
\usage{
match_args_to_func(args, func)
}
\arguments{
\item{args}{A named \code{list} containing arguments to match to `func`}

\item{func}{A function of whose argumetns to extract}
}
\description{
Helper function to extract named elements from a list to match the arguments of supplied function
}
