% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mtc_probBetter.R
\name{mtc.prob.better.table}
\alias{mtc.prob.better.table}
\title{Utility function providing pairwise probability of being better (col vs row). (Adapted from gemtc::relative.effect.table()).}
\usage{
mtc.prob.better.table(x, smaller.is.better, threshold = 0, covariate = NA)
}
\arguments{
\item{x}{Object of class \code{mtc.result} which is the output from a \code{mtc.run}.}

\item{smaller.is.better}{Logical (default is TRUE).}

\item{threshold}{Numerical value (default is 0) giving the threshold against which contrasts are being compared.}

\item{covariate}{NA; for compatibility with gemtc::relative.effect.table().}
}
\value{
An array of class `mtc.prob.better.table`
}
\description{
Utility function providing pairwise probability of being better (col vs row). (Adapted from gemtc::relative.effect.table()).
}
