% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_fit.R
\name{nma_fit}
\alias{nma_fit}
\title{Takes input data and a model plan and passes to the model engine specified. 
Current supported engines are the `gemtc` package (using mtc.model & mtc.run) or `rjags` (using jags and dic.samples functions)}
\usage{
nma_fit(model_input)
}
\arguments{
\item{model_input}{a list containing named elements fitting_data (data which has been pre-processed) and plan (list object containing all input parameters)}
}
\value{
model object of class `rjags` or `mtc.result`
}
\description{
Takes input data and a model plan and passes to the model engine specified. 
Current supported engines are the `gemtc` package (using mtc.model & mtc.run) or `rjags` (using jags and dic.samples functions)
}
