% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nma_pre_proc.R
\name{nma_pre_proc}
\alias{nma_pre_proc}
\title{NMA data pre-processing}
\usage{
nma_pre_proc(data, plan)
}
\arguments{
\item{data}{input \code{data.frame}}

\item{plan}{A \code{list} containing the model plan}
}
\value{
A network object or if gsd a list containing a network object and jags init parameters
}
\description{
NMA data pre-processing
}
\seealso{
\code{\link{groupedTTE_fp_pre_proc}}, \code{\link{groupedTTE_pwe_pre_proc}}
}
