% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan_pwe.R
\name{plan_pwe}
\alias{plan_pwe}
\title{Creates a fractional polynomial model plan}
\usage{
plan_pwe(model.pars, bth.model = c("FE", "RE"), ref.std, nma.ref.trt,
  engine = "rjags", data.type = "GSD", feprior_mean = 0,
  feprior_prec = 1e-04, descr = "Piecewise Exponential model",
  descr_s = "PWE", n.chains = NULL, n.iter = NULL, n.burnin = NULL,
  n.thin = NULL, bth.prior = NULL, rsd = 13579, model.file = NULL,
  ...)
}
\arguments{
\item{model.pars}{A named \code{list} containing `cut.pts` when using GSD}

\item{bth.model}{One of "FE" or "RE"}

\item{ref.std}{A string containing the reference study}

\item{nma.ref.trt}{A string containing the reference treatment}

\item{engine}{Only `rjags` currently supported}

\item{data.type}{Only "GSD" supported}

\item{feprior_mean}{A \code{numeric} value representing feprior_mean (default 0)}

\item{feprior_prec}{A \code{numeric} value representing feprior_prec default 0.0001}

\item{descr}{A \code{character} string to describe the model}

\item{descr_s}{Short description}

\item{n.chains}{A \code{numeric} value representing n.chains}

\item{n.iter}{A \code{numeric} value representing n.iter}

\item{n.burnin}{A \code{numeric} value representing n.burnin}

\item{n.thin}{A \code{numeric} value representing n.thin}

\item{bth.prior}{A \code{list} containing type & distr. Only required if bth.model == "FE"}

\item{rsd}{A \code{numeric} representing seed to be set}

\item{model.file}{Path to BUGS file. If NULL file name will be built from input parameters and extracted from BUGScode folder if present}

\item{...}{any other named arguments must match arguments names from \code{dic.samples} or \code{jags} functions}
}
\value{
A list containing model, engine, analysis and model params
}
\description{
Creates a fractional polynomial model plan
}
\examples{
 
model_plan <- plan_pwe(model.pars = list(cut.pts =  c(3, 10)),
                       bth.model = "FE", ref.std = "STUDY2", nma.ref.trt = "B",                        
                       n.chains = 2,
                       n.iter = 6000,
                       n.burnin = 1000,
                       n.thin = 1)
}
