% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mtc_forest.R
\name{plot_mtc_forest}
\alias{plot_mtc_forest}
\title{Utility function to do forest plot from data.frame with effect estimates.}
\usage{
plot_mtc_forest(x, sort.by = c("name", "effect")[1], lab = NULL,
  do.log = TRUE, breaks = c(0.25, 0.5, 0.8, 1, 1.25, 2, 4))
}
\arguments{
\item{x}{A data.frame with columns: Comparator, Med, CIlo, CIup.}

\item{sort.by}{Character string defining how estimates are sorted, either "name" (the default) or "effect".}

\item{lab}{Label to plot underneath x-axis.}

\item{do.log}{Logical, whether log transformed x-axis or not (default = TRUE).}

\item{breaks}{Where to put x-axis breaks and labels.}
}
\value{
A ggplot2 plot object which
}
\description{
Utility function to do forest plot from data.frame with effect estimates.
}
