% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_binary.R
\name{process_binary}
\alias{process_binary}
\title{Transforms binary data}
\usage{
process_binary(data)
}
\arguments{
\item{data}{input \code{data.frame}}
}
\value{
data.frame
}
\description{
Transforms binary data
}
