% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_gsd.R
\name{process_gsd}
\alias{process_gsd}
\title{Transforms grouped survival data}
\usage{
process_gsd(data)
}
\arguments{
\item{data}{input \code{data.frame}}
}
\value{
data.frame
}
\description{
Transforms grouped survival data
}
