% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_hr.R
\name{process_hr}
\alias{process_hr}
\title{Transforms hazard ratio data}
\usage{
process_hr(data)
}
\arguments{
\item{data}{input \code{data.frame}}
}
\value{
data.frame
}
\description{
Transforms hazard ratio data
}
