% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwe_H.R
\name{pwe_Hu}
\alias{pwe_Hu}
\title{Calculate the cumulative hazard over [0, tmax] from piecewise constant model.}
\usage{
pwe_Hu(time, cut.pts, haz.rates)
}
\arguments{
\item{time}{Time-point at which cumulative hazard evaluated, H(time) = int_0^time h(u) du.}

\item{cut.pts}{Vector of cut points.}

\item{haz.rates}{Vector of hazard rates (must have one element more than the cut.pts vector).}
}
\value{
\code{numeric} estimate of the hazard over [0, tmax]
}
\description{
Calculate the cumulative hazard over [0, tmax] from piecewise constant model.
}
