% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwe_S.R
\name{pwe_S}
\alias{pwe_S}
\title{Calculate the survivor function S(t) from a piecewise exponential model.}
\usage{
pwe_S(time, cut.pts, haz.rates)
}
\arguments{
\item{time}{Time-point at which survivor fct is evaluated, S(time) = exp(-H(time)).}

\item{cut.pts}{Vector of cut points.}

\item{haz.rates}{Vector of hazard rates (must have one element more than the cut.pts vector).}
}
\value{
\code{numeric} estimate of survivor function at time t, S(t)
}
\description{
Calculate the survivor function S(t) from a piecewise exponential model.
}
