% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_template.R
\name{create_template}
\alias{create_template}
\title{Creates a minimal project template for selected model type}
\usage{
create_template(type = c("binary", "gsd", "fp", "pwe"))
}
\arguments{
\item{type}{model type}
}
\description{
Creates a minimal project template for selected model type
}
\examples{
  create_template("binary")
}
